requirements.txt
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
pandas==2.2.3
ujson==5.10.0
boto3==1.35.81
klaviyo-api==21.0.0
google-cloud-bigquery==3.27.0